refactor: rename TaskForces.ListTaskForces to Issues.Pig
